requirements.txt
google-auth-oauthlib
gspread
flask
python-docx